app/requirements/common.txt
# Contains requirements common to all environments
Django
djangorestframework
geojson
django-cors-headers
django-cors-middleware
whitenoise